Support certificate file path
When the LND certificate expires or changes because of added hosts the
certthumbprint changes and is not updated automatically, forcing users
to update the configuration.  workaround is is to allow insecure
connections but that is obviously not secure.

This change adds a new configuration option `certfilepath` to set the
path of the certificate. The certificate located at that path is loaded
at each attempt to connect to LND and checked to contain server-supplied
certificate. If it's present the connection is accepted.

This is required for resolving
https://github.com/btcpayserver/btcpayserver/issues/2182